requirements.txt
streamlit
youtube-transcript-api
google-generativeai
PyPDF2
